45th Commit (fixing database modes and next level and level complete popup)
